refactor(deps-restorer): split the build phase into its parts
`build_modules.rs` mixed three jobs: deciding which packages may run
build scripts, running one package's scripts, and locating or repairing
the slot a build runs in. Each moves to its own module:

- `allow_build_policy` — the policy, and the dep-path key parsing it
  needs to match a lockfile entry against an `allowBuilds` pattern
- `build_one_snapshot` — one package's build, at 496 lines the largest
  single thing the file held
- `slots` — the virtual-store slot: finding it, checking it, and
  discarding a global slot whose build failed

`build_modules.rs` keeps the orchestration — options, the error type,
and the pass that walks the build sequence — at 590 lines rather than
1617.

Tests import from the module that owns what they exercise.

Pure code motion: no behaviour change, verified by comparing every
content line of the original against the union of the new modules.
